requirements.txt
streamlit>=1.28.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
groq>=0.9.0
PyMuPDF>=1.23.0
langchain-text-splitters>=0.0.1
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
